# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wealth Again API",
    "version": "0.1.0",
    "description": "Service contracts for Accounts & Ledger, Income & Tax, FX, Reporting, and Identity subsystems."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/income-tax/recalc": {
      "post": {
        "tags": ["Income & Tax"],
        "summary": "Schedule income recalculation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "taxYear": { "type": "integer" },
                  "endMonth": { "type": "integer", "minimum": 1, "maximum": 12 },
                  "startMonth": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 12
                  },
                  "cityId": { "type": "string" },
                  "userId": { "type": "string" }
                },
                "required": ["taxYear", "endMonth"]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Task scheduled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "taskId": { "type": "string" },
                    "status": { "type": "string" }
                  },
                  "required": ["taskId", "status"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/IdempotencyConflict" }
        }
      }
    },
...
    "/api/v1/reporting/dashboard": {
      "get": {
        "tags": ["Reporting"],
        "summary": "Get dashboard overview",
        "parameters": [
          {
            "name": "displayCurrency",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard totals",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/v1/reporting/income/timeseries": {
      "get": {
        "tags": ["Reporting"],
        "summary": "Get income time series",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "displayCurrency",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Income time series data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "series": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "month": { "type": "string" },
                            "value": { "type": "number" }
                          },
                          "required": ["month", "value"]
                        }
                      }
                    },
                    "summary": { "$ref": "#/components/schemas/IncomeSummary" }
                  },
                  "required": ["series", "summary"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IncomeSummary": {
        "type": "object",
        "properties": {
          "months": { "type": "integer" },
          "currency": { "type": "string" },
          "totalGross": { "type": "number" },
          "totalBonus": { "type": "number" },
          "totalLtc": { "type": "number" },
          "totalEquity": { "type": "number" },
          "totalSocialInsurance": { "type": "number" },
          "totalHousingFund": { "type": "number" },
          "totalSpecialDeductions": { "type": "number" },
          "totalTax": { "type": "number" },
          "totalNet": { "type": "number" },
          "totalIncome": { "type": "number" },
          "avgTaxRate": { "type": "number" },
          "latestTaxPaid": { "type": "number" },
          "latestTaxCumulative": { "type": "number" }
        },
        "required": [
          "months",
          "currency",
          "totalGross",
          "totalBonus",
          "totalLtc",
          "totalEquity",
          "totalSocialInsurance",
          "totalHousingFund",
          "totalSpecialDeductions",
          "totalTax",
          "totalNet",
          "totalIncome",
          "avgTaxRate",
          "latestTaxPaid",
          "latestTaxCumulative"
        ]
      },
...
      "DashboardResponse": {
        "type": "object",
        "properties": {
          "totals": {
            "type": "object",
            "properties": {
              "assets": { "type": "number" },
              "liabilities": { "type": "number" },
              "netWorth": { "type": "number" },
              "archived": { "type": "number" }
            },
            "required": ["assets", "liabilities", "netWorth", "archived"]
          },
          "displayCurrency": { "type": ["string", "null"] },
          "allocations": {
            "type": "array",
            "items": { "type": "object" }
          },
          "netWorthTrend": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "month": { "type": "string", "format": "date" },
                "netWorth": { "type": "number" },
                "assets": { "type": "number" },
                "liabilities": { "type": "number" }
              },
              "required": ["month", "netWorth"]
            }
          },
          "generatedAt": { "type": ["string", "null"], "format": "date-time" },
          "accountCount": { "type": ["integer", "null"] }
        },
        "required": ["totals", "displayCurrency", "allocations", "netWorthTrend"]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized"
      },
      "BadRequest": {
        "description": "Bad request"
      },
...
      "IdempotencyConflict": {
        "description": "Idempotency key conflict"
      }
    },
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    }
  }
}